Rename enhancer's store file
